Fixed an problem trying to store a 'size_t' value in an 'int' variable.

+++ utilities/hypre_memory.c
@@ -108,8 +108,8 @@ char *
 hypre_CAlloc( size_t count,
               size_t elt_size )
 {
-   char *ptr;
-   int   size = count*elt_size;
+   char   *ptr;
+   size_t  size = count*elt_size;
 
    if (size > 0)
    {
